overrode __init__ to build the source
