+++ moyasar/actions/create.py
@@ -3,6 +3,10 @@ import json
 
 
 class Create:
+    def __init__(self, data):
+        for key in data:
+            self.__setattr__(key, data[key])
+
     @classmethod
     def create_url(cls):
         return f'{moyasar.resource_url(cls.__name__)}'.lower()
@@ -11,6 +15,4 @@ class Create:
     def create(cls, data):
         response = moyasar.request('POST', cls.create_url(), data)
         response = json.loads(response.text)
-        resource = cls()
-        moyasar.fill_object(resource, response)
-        return resource
+        return cls(response)
