+++ moyasar/actions/fetch.py
@@ -3,6 +3,10 @@ import json
 
 
 class Fetch:
+    def __init__(self, data):
+        for key in data:
+            self.__setattr__(key, data[key])
+
     @classmethod
     def fetch_url(cls, id):
         return f'{moyasar.resource_url(cls.__name__)}/{id}'.lower()
@@ -11,6 +15,4 @@ class Fetch:
     def fetch(cls, id):
         response = moyasar.request('GET', cls.fetch_url(id), None)
         response = json.loads(response.text)
-        resource = cls()
-        moyasar.fill_object(resource, response)
-        return resource
+        return cls(response)
