+++ moyasar/actions/list.py
@@ -3,6 +3,10 @@ import json
 
 
 class List:
+    def __init__(self, data):
+        for key in data:
+            self.__setattr__(key, data[key])
+
     @classmethod
     def list_url(cls):
         return f'{moyasar.resource_url(cls.__name__)}'.lower()
@@ -14,8 +18,6 @@ class List:
         field_name = f'{cls.__name__}s'.lower()
         resource_list = []
         for resource in response[field_name]:
-            r = cls()
-            moyasar.fill_object(r, resource)
-            resource_list.append(r)
+            resource_list.append(cls(resource))
 
         return resource_list
